requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
phonenumbers==8.12.56
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
phonenumbers==8.12.56
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
phonenumbers==8.12.56
PyJWT==2.5.0
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
